Conflicts:
	flexodesktop/GUI/flexographicalengine/pom.xml
	flexodesktop/GUI/flexographicalengine/src/main/java/org/openflexo/fge/ConnectorGraphicalRepresentation.java
	flexodesktop/GUI/flexographicalengine/src/main/java/org/openflexo/fge/ShapeGraphicalRepresentation.java
	flexodesktop/GUI/flexographicalengine/src/main/java/org/openflexo/fge/connectors/LineConnector.java
	flexodesktop/GUI/flexographicalengine/src/main/java/org/openflexo/fge/connectors/rpc/RectPolylinConnector.java
	flexodesktop/GUI/flexographicalengine/src/main/java/org/openflexo/fge/view/FGELayeredView.java
	flexodesktop/GUI/flexographicalengine/src/main/java/org/openflexo/fge/view/LabelView.java
	flexodesktop/GUI/flexographicalengine/src/main/java/org/openflexo/fge/view/ShapeView.java
	flexodesktop/GUI/flexographicalengine/src/main/java/org/openflexo/fge/view/listener/FGEViewMouseListener.java
	flexodesktop/GUI/flexographicalengine/src/main/java/org/openflexo/fge/view/listener/FocusRetriever.java
	flexodesktop/model/flexofoundation/src/main/java/org/openflexo/foundation/view/diagram/viewpoint/action/AbstractDeclareShapeInEditionPattern.java
	flexodesktop/model/flexofoundation/src/main/java/org/openflexo/foundation/view/diagram/viewpoint/action/AbstractPushEditonPatternToPalette.java
	flexodesktop/modules/flexovieweditor/src/main/java/org/openflexo/ve/diagram/DiagramShapeGR.java
	flexodesktop/modules/flexovieweditor/src/main/java/org/openflexo/ve/diagram/DrawEdgeControl.java
	pom.xml

+++ flexodesktop/model/flexofoundation/src/main/java/org/openflexo/foundation/wkf/node/AbstractActivityNode.java
@@ -265,7 +265,9 @@ public abstract class AbstractActivityNode extends FatherNode implements Metrics
 
 								@Override
 								public void propertyChange(PropertyChangeEvent evt) {
-									manager.delete();
+									if (manager != null) {
+										manager.delete();
+									}
 									manager = null;
 									if (role != null && role.getObject(true) != null) {
 										observedRole = role.getObject(true);
@@ -930,7 +932,7 @@ public abstract class AbstractActivityNode extends FatherNode implements Metrics
 
 	@Override
 	public void objectDeleted(FlexoModelObjectReference<?> reference) {
-		if (!getProcess().getFlexoResource().isConverting()) {
+		if (getProcess() != null && getProcess().getFlexoResource() != null && !getProcess().getFlexoResource().isConverting()) {
 			if (role == reference) {
 				setRole(null);
 			} else if (roleA == reference) {
